Revert "Modern dashboard"
