--- Client/Enums.cs
@@ -1,12 +1,0 @@
-﻿namespace BlazorApp.Client
-{
-    public enum DashboardCardType
-    {
-        Temp,
-        Humidity,
-        Dew,
-        Press,
-        Wind,
-        Rain
-    }
-}
