+++ Shared/UnitConversions.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BlazorApp.Shared
+{
+    public static class UnitConversions
+    {
+        public static float celsiusToFahrenheit(float temperatureC)
+        {
+            throw new NotImplementedException();
+        }
+
+        public static float celsiusToKelvin(float temperatureC)
+        {
+            throw new NotImplementedException();
+        }
+
+        public static float mmToInch(float rainMM)
+        {
+            throw new NotImplementedException();
+        }
+
+        public static float kmhToMph(float windKph)
+        {
+            throw new NotImplementedException();
+        }
+
+        public static float kmhToKnot(float windKph)
+        {
+            throw new NotImplementedException();
+        }
+
+        public static float hPaToInHg(float pressureHpa)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
